Added grouping field compose primitive. Split off preprocessing pipeline and fully specified pipelines.
